simplify: extract constantTimeEqual + drop FTS type gymnastics + memoize buildPermissions
Three simplify wins from the internal review:

1. **Shared constantTimeEqual** (src/utils/crypto.ts) — extracted the
   duplicate `tokenMatches` (http.ts) and `safeEqual` (oauth-handlers.ts).
   Same padding trick (timingSafeEqual against a zero buffer when lengths
   differ, avoids length-leak). verifyPkceS256 also routes through it now,
   dropping the last duplicate of the buffer-eq dance.

2. **FTS transaction() cast** (fts-service.ts) — tightened the local
   better-sqlite3 shim to `F extends (...args: never[]) => unknown`, which
   preserves the inner fn's argument types. Drops the triple `as unknown as`
   ladder; the block is now seven lines of concrete types instead of cast
   soup.

3. **buildPermissions memoization** (grant-manager.ts) — the gate called
   buildPermissions twice per check (once for the override path, once for
   the preset path). Preset→permissions is pure, so a per-manager Map
   cache eliminates the duplicate materialization. Also removed the
   identical `presetAllows` wrapper (it just called globalAllows).

Not included (deliberately): GrantStore.revoke() → deny() dedup. They
share an implementation but express different semantic intents to
readers (active → revoked vs pending → denied); worth the 3 LOC.

No behavior changes. All 1564 tests still pass.
